+++ resources/js/components/DetailTabs.vue
@@ -19,7 +19,7 @@
                 aria-label="Tabs"
                 class="tab-menu"
             >
-              <Button
+              <a
                   v-for="(tab, key) in getSortedTabs(tabs)"
                   :key="key"
                   :dusk="tab.slug + '-tab'"
@@ -38,7 +38,7 @@
                     aria-hidden="true"
                     class="tabs-bg-transparent tabs-absolute tabs-inset-x-0 tabs-bottom-0 tabs-h-0.5"
                 ></span>
-              </Button>
+              </a>
             </nav>
           </div>
         </div>
@@ -77,209 +77,19 @@
 
 <script>
 import BehavesAsPanel from '../../../vendor/laravel/nova/resources/js/mixins/BehavesAsPanel';
+import HasTabs from "../mixins/HasTabs";
+
 import Heading from '../../../vendor/laravel/nova/resources/js/components/Heading.vue';
 import Card from '../../../vendor/laravel/nova/resources/js/components/Card.vue';
-import orderBy from 'lodash/orderBy';
-import { parseLocationHash, updateLocationHash } from '../utils/hash.js';
 
 export default {
-  mixins: [BehavesAsPanel],
-
+  mixins: [BehavesAsPanel, HasTabs],
   components: {Card, Heading},
-
   props: {
-    panel: {
-      type: Object,
-      required: true,
-    },
-
-    name: {
-      default: 'Panel',
-    },
-
     mode: {
       type: String,
       default: 'detail',
     },
-
-    fields: {
-      type: Array,
-      default: [],
-    },
-
-    resourceName: {
-      type: String,
-      required: true,
-    },
-
-    resourceId: {
-      type: [Number, String],
-    },
-
-    relatedResourceName: {
-      type: String,
-    },
-
-    relatedResourceId: {
-      type: [Number, String],
-    },
-
-    viaResource: {
-      type: String,
-    },
-
-    viaResourceId: {
-      type: [Number, String],
-    },
-
-    viaRelationship: {
-      type: String,
-    },
-  },
-
-  data() {
-    return {
-      tabs: null,
-      selectedTab: {},
-      darkModeClass: ''
-    };
-  },
-
-  /**
-   * Get the tabs and their respective fields when mounted
-   * and show the first tab by default.
-   */
-  mounted() {
-    this.observer = new MutationObserver(mutations => {
-      for (const m of mutations) {
-        const newValue = m.target.getAttribute(m.attributeName);
-        this.$nextTick(() => {
-          this.darkModeClass = newValue.includes('dark') ? 'tabs-dark' : ''
-        });
-      }
-    });
-
-    this.observer.observe(document.documentElement, {
-      attributes: true,
-      attributeOldValue : true,
-      attributeFilter: ['class'],
-    });
-
-    this.darkModeClass = document.documentElement.classList.contains('dark') ? 'tabs-dark' : '';
-
-    const tabs = this.tabs = this.panel.fields.reduce((tabs, field) => {
-      if (!(field.tabSlug in tabs)) {
-        tabs[field.tabSlug] = {
-          name: field.tab,
-          slug: field.tabSlug,
-          position: field.tabPosition,
-          init: false,
-          listable: field.listableTab,
-          fields: [],
-          properties: field.tabInfo,
-          classes: 'fields-tab',
-        };
-        if (['belongs-to-many-field', 'has-many-field', 'has-many-through-field', 'has-one-through-field', 'morph-to-many-field',].includes(field.component)) {
-          tabs[field.tabSlug].classes = 'relationship-tab';
-        }
-      }
-
-      tabs[field.tabSlug].fields.push(field);
-
-      return tabs;
-    }, {});
-
-    const routeTabs = parseLocationHash()
-    const currentTabSlug = routeTabs[this.panel.name]
-    if (tabs[currentTabSlug]) {
-      this.handleTabClick(tabs[currentTabSlug])
-    } else {
-      this.handleTabClick(tabs[Object.keys(tabs)[0]], true);
-    }
-  },
-
-  methods: {
-
-    /**
-     * Handle tabs being clicked
-     *
-     * @param tab
-     * @param updateUri
-     */
-    handleTabClick(tab, updateUri = true) {
-      this.selectedTab = tab;
-      if (updateUri) {
-        this.updateHash()
-      }
-    },
-
-    /**
-     * Update the location hash to persist route state
-     *
-     * @param tab
-     * @param updateUri
-     */
-    updateHash() {
-      const routeTabs = parseLocationHash()
-      routeTabs[this.panel.name] = this.selectedTab.slug
-      updateLocationHash(routeTabs)
-    },
-
-    /**
-     * Get the component name.
-     *
-     * @param field
-     * @returns {string}
-     */
-    getComponentName(field) {
-      return field.prefixComponent
-          ? 'detail-' + field.component
-          : field.component
-    },
-
-    /**
-     * Get body class for tabbed field panel
-     *
-     * @param tab
-     * @returns {string}
-     */
-    getBodyClass(tab) {
-      return tab.properties.bodyClass;
-    },
-
-    /**
-     * Get reference name for tab
-     *
-     * @param tab
-     * @returns {string}
-     */
-    getTabRefName(tab) {
-      return `tab-${tab.slug}`;
-    },
-
-    /**
-     * Check if the specified tab is the current opened one
-     *
-     * @param tab
-     * @returns {boolean}
-     */
-    getIsTabCurrent(tab) {
-      return this.selectedTab === tab || (!this.selectedTab && this.tabs[Object.keys(this.tabs)[0]] === tab)
-    },
-
-    /**
-     * Sort the tabs object by their respective positions using lodash
-     *
-     * @param tabs
-     * @returns {object}
-     */
-    getSortedTabs(tabs) {
-      return orderBy(tabs, [c => c.position], ['asc']);
-    }
-
-  },
-  beforeDestroy() {
-    this.observer.disconnect();
-  },
+  }
 };
 </script>
